Trying to remove jersey in favor of default spring boot web.
