+++ src/main/java/com/tibiawiki/controller/KeysResource.java
@@ -1,0 +1,74 @@
+package com.tibiawiki.controller;
+
+import com.tibiawiki.domain.objects.Key;
+import com.tibiawiki.domain.objects.WikiObject;
+import com.tibiawiki.domain.objects.validation.ValidationException;
+import com.tibiawiki.process.ModifyAny;
+import com.tibiawiki.process.RetrieveKeys;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import org.json.JSONObject;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@Api(value = "Keys")
+@RequestMapping("/keys")
+@RequiredArgsConstructor
+public class KeysResource {
+
+    private final RetrieveKeys retrieveKeys;
+    private final ModifyAny modifyAny;
+
+    @GetMapping(value = "")
+    @ApiOperation(value = "Get a list of keys")
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "list of keys retrieved")
+    })
+    public ResponseEntity<Object> getKeys(@ApiParam(value = "optionally expands the result to retrieve not only " +
+            "the key names but the full keys", required = false)
+                                          @RequestParam(value = "expand", required = false) Boolean expand) {
+        return ResponseEntity.ok()
+                .body(expand != null && expand
+                        ? retrieveKeys.getKeysJSON().map(JSONObject::toMap)
+                        : retrieveKeys.getKeysList()
+                );
+    }
+
+    @GetMapping("/{name}")
+    @ApiOperation(value = "Get a specific key by name")
+    public ResponseEntity<String> getKeysByName(@PathVariable("name") String name) {
+        return retrieveKeys.getKeyJSON(name)
+                .map(a -> ResponseEntity.ok()
+                        .body(a.toString(2)))
+                .orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
+    @PutMapping(value = "")
+    @ApiOperation(value = "Modify a key")
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "the changed key"),
+            @ApiResponse(code = 400, message = "the provided changed key is not valid"),
+            @ApiResponse(code = 401, message = "not authorized to edit without providing credentials")
+    })
+    public ResponseEntity<WikiObject> putKey(@RequestBody Key key, @RequestHeader("X-WIKI-Edit-Summary") String editSummary) {
+        return modifyAny.modify(key, editSummary)
+                .map(a -> ResponseEntity.ok()
+                        .body(a))
+                .recover(ValidationException.class, e -> ResponseEntity.badRequest().build())
+                .recover(e -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build())
+                .get();
+    }
+}
