+++ src/main/java/com/tibiawiki/controller/LocationsResource.java
@@ -1,0 +1,74 @@
+package com.tibiawiki.controller;
+
+import com.tibiawiki.domain.objects.Location;
+import com.tibiawiki.domain.objects.WikiObject;
+import com.tibiawiki.domain.objects.validation.ValidationException;
+import com.tibiawiki.process.ModifyAny;
+import com.tibiawiki.process.RetrieveLocations;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import org.json.JSONObject;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@Api(value = "Locations")
+@RequestMapping("/locations")
+@RequiredArgsConstructor
+public class LocationsResource {
+
+    private final RetrieveLocations retrieveLocations;
+    private final ModifyAny modifyAny;
+
+    @GetMapping(value = "")
+    @ApiOperation(value = "Get a list of locations")
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "list of locations retrieved")
+    })
+    public ResponseEntity<Object> getLocations(@ApiParam(value = "optionally expands the result to retrieve not only " +
+            "the location names but the full locations", required = false)
+                                               @RequestParam(value = "expand", required = false) Boolean expand) {
+        return ResponseEntity.ok()
+                .body(expand != null && expand
+                        ? retrieveLocations.getLocationsJSON().map(JSONObject::toMap)
+                        : retrieveLocations.getLocationsList()
+                );
+    }
+
+    @GetMapping("/{name}")
+    @ApiOperation(value = "Get a specific location by name")
+    public ResponseEntity<String> getLocationsByName(@PathVariable("name") String name) {
+        return retrieveLocations.getLocationJSON(name)
+                .map(a -> ResponseEntity.ok()
+                        .body(a.toString(2)))
+                .orElseGet(() -> ResponseEntity.notFound().build());
+    }
+
+    @PutMapping(value = "")
+    @ApiOperation(value = "Modify a location")
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "the changed location"),
+            @ApiResponse(code = 400, message = "the provided changed location is not valid"),
+            @ApiResponse(code = 401, message = "not authorized to edit without providing credentials")
+    })
+    public ResponseEntity<WikiObject> putLocation(@RequestBody Location location, @RequestHeader("X-WIKI-Edit-Summary") String editSummary) {
+        return modifyAny.modify(location, editSummary)
+                .map(a -> ResponseEntity.ok()
+                        .body(a))
+                .recover(ValidationException.class, e -> ResponseEntity.badRequest().build())
+                .recover(e -> ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build())
+                .get();
+    }
+}
