--- src/main/java/com/tibiawiki/serviceinterface/CreaturesResource.java
@@ -1,89 +1,0 @@
-package com.tibiawiki.serviceinterface;
-
-import com.tibiawiki.domain.objects.Creature;
-import com.tibiawiki.domain.objects.validation.ValidationException;
-import com.tibiawiki.process.ModifyAny;
-import com.tibiawiki.process.RetrieveCreatures;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.HeaderParam;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
-@Component
-@Api(value = "Creatures")
-@Path("/creatures")
-public class CreaturesResource {
-
-    private RetrieveCreatures retrieveCreatures;
-    private ModifyAny modifyAny;
-
-    @Autowired
-    private CreaturesResource(RetrieveCreatures retrieveCreatures, ModifyAny modifyAny) {
-        this.retrieveCreatures = retrieveCreatures;
-        this.modifyAny = modifyAny;
-    }
-
-    @GET
-    @ApiOperation(value = "Get a list of creatures")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "list of creatures retrieved")
-    })
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response getCreatures(@ApiParam(value = "optionally expands the result to retrieve not only " +
-            "the creature names but the full creatures", required = false)
-                                 @QueryParam("expand") Boolean expand) {
-        return Response.ok()
-                .entity(expand != null && expand
-                        ? retrieveCreatures.getCreaturesJSON().map(JSONObject::toMap)
-                        : retrieveCreatures.getCreaturesList()
-                )
-                .build();
-    }
-
-    @GET
-    @Path("/{name}")
-    @ApiOperation(value = "Get a specific creature by name")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response getCreatureByName(@PathParam("name") String name) {
-        return retrieveCreatures.getCreatureJSON(name)
-                .map(a -> Response.ok()
-                        .entity(a.toString(2))
-                        .build())
-                .orElseGet(() -> Response.status(Response.Status.NOT_FOUND)
-                        .build());
-    }
-
-    @PUT
-    @ApiOperation(value = "Modify a creature")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "the changed creature"),
-            @ApiResponse(code = 400, message = "the provided changed creature is not valid"),
-            @ApiResponse(code = 401, message = "not authorized to edit without providing credentials")
-    })
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response putCreature(Creature creature, @HeaderParam("X-WIKI-Edit-Summary") String editSummary) {
-        return modifyAny.modify(creature, editSummary)
-                .map(a -> Response.ok()
-                        .entity(a)
-                        .build())
-                .recover(ValidationException.class, e -> Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build())
-                .recover(e -> Response.serverError().build())
-                .get();
-    }
-}
