--- src/main/java/com/tibiawiki/serviceinterface/LootStatisticsResource.java
@@ -1,58 +1,0 @@
-package com.tibiawiki.serviceinterface;
-
-import com.tibiawiki.process.RetrieveLoot;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import lombok.RequiredArgsConstructor;
-import org.json.JSONObject;
-import org.springframework.stereotype.Component;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
-@Component
-@Api(value = "Loot Statistics")
-@Path("/loot")
-@RequiredArgsConstructor
-public class LootStatisticsResource {
-
-    private final RetrieveLoot retrieveLoot;
-
-    @GET
-    @ApiOperation(value = "Get a list of loot statistics")
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "list of loot statistics retrieved")
-    })
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response getLoot(@ApiParam(value = "optionally expands the result to retrieve not only " +
-            "the loot statistics page names but the full loot statistics", required = false)
-                            @QueryParam("expand") Boolean expand) {
-        return Response.ok()
-                .entity(expand != null && expand
-                        ? retrieveLoot.getLootJSONObject().map(JSONObject::toMap)
-                        : retrieveLoot.getLootList()
-                )
-                .build();
-    }
-
-    @GET
-    @Path("/{name}")
-    @ApiOperation(value = "Get a specific loot statistics page by creature name")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response getLootByName(@PathParam("name") String name) {
-        return retrieveLoot.getLootJSONObject("Loot_Statistics:" + name)
-                .map(a -> Response.ok()
-                        .entity(a.toString(2))
-                        .build())
-                .orElseGet(() -> Response.status(Response.Status.NOT_FOUND)
-                        .build());
-    }
-}
