--- src/main/java/com/tibiawiki/serviceinterface/config/CORSResponseFilter.java
@@ -1,18 +1,0 @@
-package com.tibiawiki.serviceinterface.config;
-
-import javax.ws.rs.container.ContainerRequestContext;
-import javax.ws.rs.container.ContainerResponseContext;
-import javax.ws.rs.container.ContainerResponseFilter;
-import javax.ws.rs.core.MultivaluedMap;
-import java.io.IOException;
-
-public class CORSResponseFilter implements ContainerResponseFilter {
-
-    public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) throws IOException {
-
-        MultivaluedMap<String, Object> headers = responseContext.getHeaders();
-        headers.add("Access-Control-Allow-Origin", "*");
-        headers.add("Access-Control-Allow-Methods", "GET, POST, DELETE, PUT");
-        headers.add("Access-Control-Allow-Headers", "X-Requested-With, Content-Type");
-    }
-}
